feature: multiple configs possible with conditions
Extend the plugin to allow condition blocks in configuration blocks. All
configuration blocks that would apply are now additionally scanned for
their conditions. The more conditions are present and apply, the higher
the priority of the corresponding configuration block. In the end, the
configuration block with the highest priority is chosen.