fix: configs with * project were not evaluated
When a config section matches for all projects, it may not be processed
at all, because the matching against config sections with a name
supplied returns zero matches, but no exception. Currently, only
exceptions in the first match lead to an evaluation of * project
matches. This issue needs to be investigated further.